Add change password functionality and video upload endpoint
- Introduced a new `ChangePasswordRequest` model and corresponding `/change-password` endpoint to allow users to change their passwords after verifying their current password.
- Enhanced the file upload service by adding a new `/upload-video` endpoint for video uploads, including MIME type validation and size restrictions.
- Updated the file service to support video uploads, ensuring proper handling and storage of video files.
- Modified existing image upload logic to improve content type inference based on file extensions, enhancing overall file handling capabilities.

+++ backend/app/services/comment_service.py
@@ -321,7 +321,7 @@ class CommentService:
         result = self.db.table("show_image_reviews").select("*").eq("image_id", image_id).order("created_at", desc=True).execute()
         return [self._format_image_review(r) for r in result.data or []]
 
-    def create_image_review(self, image_id: int, user_id: int, rating: int, content: str) -> Optional[ImageReview]:
+    def create_image_review(self, image_id: int, user_id: int, rating: float, content: str) -> Optional[ImageReview]:
         """创建秀场图片评论"""
         result = self.db.table("show_image_reviews").insert({
             "image_id": image_id,
